+++ Backend/Core/Sync.h
@@ -86,7 +86,7 @@ public:
     QJsonObject lastSettingsResponseData() const;
 
     Q_INVOKABLE void pushSensorTemperatureC(const double &temperatureC);
-    Q_INVOKABLE void pushSensorHumidity(const double &humidity);
+    Q_INVOKABLE void pushSensorHumidity(const int &humidity);
     Q_INVOKABLE void pushSensorCo2ID(const int &co2ID);
 
     Q_INVOKABLE void updateAddressInformationManual(const QVariantMap &data);
@@ -162,8 +162,8 @@ signals:
     void resetFactoryFinished(bool ok, const QString &message = "");
 
     void sensorTemperaturePushed(bool success, double pushedTemperatureC);
-    void sensorHumidityPushed(bool success, double pushedHumidity);
-    void sensorCo2Pushed(bool success, double pushedCo2Id);
+    void sensorHumidityPushed(bool success, int pushedHumidity);
+    void sensorCo2Pushed(bool success, int pushedCo2Id);
 
     void clientAddressUpdatingFinished(bool success, const QString &err = "", bool needToRetry = false);
